phy: Enable physics with affine plane transformations (1/x?)
This series of commits will enable physics even if affine
transformations have been used on the plane.

So far the plane did not have its own model matrix (other than an
implicit identity matrix) because this made looking up the triangle
under the sphere trivial. The drawback was that this made it
impossible to move or scale the plane in space without breaking the
physics.

As a first step this commit adds an explicit model matrix to the
phyPlane.