phy: Enable physics with affine plane transformations (2/x?)
Bouncing of the plane is functional now when affine transformations
have been applied to the plane. The terraining_testing program
illustrates this by slowly tilting the plane after frame 500.

FIXME: As can be seen, spheres in the upper half of the plane, where
the plane tilts towards them don't get pushed up by the plane yet. For
the other half where the plane moves downwards, away from the spheres,
this problem does not occur.